Timing bug fix
Fixed a bug that messed up the timing when any event set the timing to 0. When the timing is directly set to 0, then the next event will be executed immediately.
Before the bugfix, the code didn't expect a timing set to 0, this caused an underflow. Now the code that decrements the counter is completely skipped instead.