FM Bugfix
When playing a note on a precedently reset FM channel, a click could be heard. That was caused by the fact that the correct OP TLs
were set slightly after the note was played. This has been fixed, and now before an FM note is played, the driver checks whether or
not a TL update was requested, and if that's the case executes it.